Optimisations
- Images for the image carousel is now embedded into the launcher instead of downloading from the web. Since they are updated rarely, this is much more reliable and optimal
- New images for image carousel

+++ H1EmuLauncher/Windows/Launcher.xaml.cs
@@ -448,11 +448,8 @@ namespace H1EmuLauncher
             if (!File.Exists($"{Info.APPLICATION_DATA_PATH}\\H1EmuLauncher\\args.txt"))
                 File.Create($"{Info.APPLICATION_DATA_PATH}\\H1EmuLauncher\\args.txt");
 
-            try
-            {
+            if (Directory.Exists($"{Info.APPLICATION_DATA_PATH}\\H1EmuLauncher\\CarouselImages"))
                 Directory.Delete($"{Info.APPLICATION_DATA_PATH}\\H1EmuLauncher\\CarouselImages", true);
-            }
-            catch { }
 
             argsWatcher.Path = $"{Info.APPLICATION_DATA_PATH}\\H1EmuLauncher";
             argsWatcher.Filter = "args.txt";
